Reduce number of cfg for race
